Event Processor API - Moved fake classes in test.
